refactor: KEEP-356 group workflow editor files under lib/workflow/editor
Authoring helpers (auto-layout, edge utils, output field derivation,
template helpers, extension registry, vote/sanitize logic, show-when)
were spread across lib/ root and lib/workflow/. Group them in
lib/workflow/editor/ to mirror the executor + nodes structure.

Moves from lib/ root:
- auto-layout.ts            -> lib/workflow/editor/auto-layout.ts
- edge-handle-utils.ts      -> lib/workflow/editor/edge-handle-utils.ts
- action-output-fields.ts   -> lib/workflow/editor/action-output-fields.ts
- trigger-output-fields.ts  -> lib/workflow/editor/trigger-output-fields.ts
- builtin-variables.ts      -> lib/workflow/editor/builtin-variables.ts
- template-helpers.ts       -> lib/workflow/editor/template-helpers.ts
- template-utils.ts         -> lib/workflow/editor/template-utils.ts
- extension-registry.ts     -> lib/workflow/editor/extension-registry.ts
- extensions.tsx            -> lib/workflow/editor/extensions.tsx

Moves from lib/workflow/ (no longer mixed with structural files):
- edge-helpers.ts           -> lib/workflow/editor/edge-helpers.ts
- sanitize-nodes.ts         -> lib/workflow/editor/sanitize-nodes.ts
- show-when.ts              -> lib/workflow/editor/show-when.ts
- vote-rate-limit.ts        -> lib/workflow/editor/vote-rate-limit.ts
- votes.ts                  -> lib/workflow/editor/votes.ts
