+++ components/ui/sql-template-editor.tsx
@@ -14,7 +14,7 @@ import {
   getCommonFields,
   getNodeDisplayName,
   sanitizeNodeId,
-} from "@/lib/template-helpers";
+} from "@/lib/workflow/editor/template-helpers";
 import { useStableRef } from "@/lib/use-stable-ref";
 import { getAvailableFields, type NodeOutputs } from "@/lib/utils/template";
 import {
