+++ components/ui/template-badge-input.tsx
@@ -2,7 +2,7 @@
 
 import { useAtom } from "jotai";
 import { useEffect, useRef, useState } from "react";
-import { doesNodeExist, getDisplayTextForTemplate } from "@/lib/template-utils";
+import { doesNodeExist, getDisplayTextForTemplate } from "@/lib/workflow/editor/template-utils";
 import { cn } from "@/lib/utils";
 import { nodesAtom, selectedNodeAtom } from "@/lib/workflow/store";
 import { TemplateAutocomplete } from "./template-autocomplete";
