+++ lib/workflow/editor/extension-registry.ts
