+++ lib/workflow/store.ts
@@ -1,8 +1,8 @@
 import type { Edge, EdgeChange, Node, NodeChange } from "@xyflow/react";
 import { applyEdgeChanges, applyNodeChanges } from "@xyflow/react";
 import { atom } from "jotai";
-import { computeAutoLayout } from "@/lib/auto-layout";
-import { buildExecutionLogsMap } from "@/lib/template-helpers";
+import { computeAutoLayout } from "@/lib/workflow/editor/auto-layout";
+import { buildExecutionLogsMap } from "@/lib/workflow/editor/template-helpers";
 import { api } from "@/lib/api-client";
 
 export type WorkflowNodeType = "trigger" | "action" | "add";
