+++ tests/unit/builtin-variables.test.ts
@@ -8,7 +8,7 @@ import {
   BUILTIN_NODE_LABEL,
   BUILTIN_VARIABLE_FIELDS,
   getBuiltinVariables,
-} from "@/lib/builtin-variables";
+} from "@/lib/workflow/editor/builtin-variables";
 import { evaluateConditionExpression } from "@/lib/workflow/executor/executor.workflow";
 
 describe("builtin-variables", () => {
