+++ tests/unit/condition-executor-handles.test.ts
@@ -2,7 +2,7 @@ import { describe, expect, it, vi } from "vitest";
 
 vi.mock("server-only", () => ({}));
 
-import { buildEdgesBySourceHandle } from "@/lib/edge-handle-utils";
+import { buildEdgesBySourceHandle } from "@/lib/workflow/editor/edge-handle-utils";
 import { identifyLoopBody } from "@/lib/workflow/executor/executor.workflow";
 import type { WorkflowNode } from "@/lib/workflow/store";
