+++ tests/unit/condition-foreach-multi-loop.test.ts
@@ -16,7 +16,7 @@ import { describe, expect, it, vi } from "vitest";
 
 vi.mock("server-only", () => ({}));
 
-import { buildEdgesBySourceHandle } from "@/lib/edge-handle-utils";
+import { buildEdgesBySourceHandle } from "@/lib/workflow/editor/edge-handle-utils";
 import {
   identifyLoopBody,
   resolveBodyConditionTargets,
