+++ tests/unit/convergence-barrier.test.ts
@@ -9,7 +9,7 @@ import {
   propagateConvergenceSkips,
   signalConvergenceArrival,
 } from "@/lib/workflow/executor/convergence-barrier";
-import { buildEdgesBySourceHandle } from "@/lib/edge-handle-utils";
+import { buildEdgesBySourceHandle } from "@/lib/workflow/editor/edge-handle-utils";
 
 describe("convergence barrier", () => {
   describe("basic convergence: A -> [B, C, D] -> E", () => {
